pyproject.toml
[tool.poetry]
name = "chartocr"
version = "0.1.0"
description = "OCR Web Application"
authors = ["Replit User"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.0.1"
pillow = "^8.3.2"
pytesseract = "^0.3.8"
flask-cors = "^3.0.10"
python-dotenv = "^0.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"